Fix #16394: Added super-admin-utils class for puppeteer-E2E-acceptance-tests (#17225)
* added super-admin-utils class

* addressed comment by Sean

* addressed Sean comment about Pop up dialog

* removed learner-dashboard constant

* changed class name ( puppeteerUtilities ->  baseUser )

* added commented link for event 'dialog'

* addressed Sean comments

---------
